+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Support for logging saved search added
 - OpenAPI Support in API Gateway
 - Support for Autonomous Database with Dedicated Infrastructure - Autonomous Data Guard
+- Support for `Resource Discovery` added for `Logging` service
 
 ## 3.96.0 (October 07, 2020)